Make private connect flow fully catalog-driven
Replace the hardcoded Schulnetz/OdAOrg private connect screens with one generic screen that renders the chosen system's loginFields (DynamicLoginForm) and branches only on the backend-provided loginMethod. PrivateAccount is keyed on systemKey + loginMethod; the data fetch routes by loginMethod. No provider or login fields are hardcoded in the client.
