+++ lib/ui/private/private_connect_flow.dart
@@ -2,12 +2,11 @@ import 'package:flutter/material.dart';
 
 import '../../services/school_systems_service.dart';
 import '../dashboard/widgets/add_school_modal.dart';
-import '../odaorg/private_odaorg_connect_screen.dart';
-import '../schulnetz/private_connect_screen.dart';
+import 'private_connect_screen.dart';
 
 /// Private-mode connect flow: show the backend's school-system picker, then run
-/// the chosen system's private connect screen. Returns true if a connection was
-/// stored on-device.
+/// the generic connect screen for the chosen system (driven entirely by the
+/// catalog descriptor). Returns true if a connection was stored on-device.
 Future<bool> runPrivateConnectFlow(BuildContext context) async {
   final systems = await SchoolSystemsService.fetch();
   if (!context.mounted) return false;
@@ -15,10 +14,9 @@ Future<bool> runPrivateConnectFlow(BuildContext context) async {
   final key = await showAddSchoolModal(context, systems);
   if (key == null || !context.mounted) return false;
 
-  final Widget screen = key == 'odaorg'
-      ? const OdaorgPrivateConnectScreen()
-      : const PrivateConnectScreen();
-  final ok = await Navigator.of(context)
-      .push<bool>(MaterialPageRoute(builder: (_) => screen));
+  final system = systems.firstWhere((s) => s.key == key);
+  final ok = await Navigator.of(context).push<bool>(
+    MaterialPageRoute(builder: (_) => PrivateConnectScreen(system: system)),
+  );
   return ok == true;
 }
